+++ experimental/packages/opentelemetry-exporter-metrics-otlp-http/src/OTLPMetricExporterBase.ts
@@ -16,14 +16,18 @@
 
 import { ExportResult, getEnv } from '@opentelemetry/core';
 import {
+  Aggregation,
+  AggregationSelector,
   AggregationTemporality,
   AggregationTemporalitySelector,
+  DefaultAggregationSelector,
   InstrumentType,
   PushMetricExporter,
   ResourceMetrics,
 } from '@opentelemetry/sdk-metrics';
 import {
   AggregationTemporalityPreference,
+  HistogramAggregationPreference,
   OTLPMetricExporterOptions,
 } from './OTLPMetricExporterOptions';
 import { OTLPExporterBase } from '@opentelemetry/otlp-exporter-base';
@@ -104,6 +108,67 @@ function chooseTemporalitySelector(
   return chooseTemporalitySelectorFromEnvironment();
 }
 
+function makeCompoundAggregatorSelector(
+  aggregationSelector: AggregationSelector,
+  instrumentType: InstrumentType,
+  aggregation: Aggregation
+): AggregationSelector {
+  return (_instrumentType: InstrumentType) => {
+    if (_instrumentType === instrumentType) {
+      return aggregation;
+    }
+    return aggregationSelector(_instrumentType);
+  };
+}
+
+export const ExplicitHistogramAggregationSelector =
+  makeCompoundAggregatorSelector(
+    DefaultAggregationSelector,
+    InstrumentType.HISTOGRAM,
+    Aggregation.Histogram()
+  );
+
+export const ExponentialHistogramAggregationSelector =
+  makeCompoundAggregatorSelector(
+    DefaultAggregationSelector,
+    InstrumentType.HISTOGRAM,
+    Aggregation.ExponentialHistogram()
+  );
+
+function chooseHistogramAggregationFromEnvironment(): AggregationSelector {
+  const env = getEnv();
+  const configuredHistogramAggregation =
+    env.OTEL_EXPORTER_OTLP_METRICS_DEFAULT_HISTOGRAM_AGGREGATION.trim().toLowerCase();
+
+  if (configuredHistogramAggregation === 'base2_exponential_bucket_histogram') {
+    return ExponentialHistogramAggregationSelector;
+  }
+  if (configuredHistogramAggregation === 'explicit_bucket_histogram') {
+    return ExplicitHistogramAggregationSelector;
+  }
+
+  diag.warn(
+    `OTEL_EXPORTER_OTLP_METRICS_DEFAULT_HISTOGRAM_AGGREGATION is set to '${env.OTEL_EXPORTER_OTLP_METRICS_DEFAULT_HISTOGRAM_AGGREGATION}', but only 'explicit_bucket_histogram' and 'base2_exponential_bucket_histogram' are allowed. Using default ('explicit_bucket_histogram') instead.`
+  );
+  return ExplicitHistogramAggregationSelector;
+}
+
+function chooseHistogramAggregation(
+  histogramPreference?: HistogramAggregationPreference
+): AggregationSelector {
+  // Directly passed preference has priority.
+  if (histogramPreference != null) {
+    if (
+      histogramPreference ===
+      HistogramAggregationPreference.EXPONENTIAL_BUCKET_HISTOGRAM
+    ) {
+      return ExponentialHistogramAggregationSelector;
+    }
+    return ExplicitHistogramAggregationSelector;
+  }
+  return chooseHistogramAggregationFromEnvironment();
+}
+
 export class OTLPMetricExporterBase<
   T extends OTLPExporterBase<
     OTLPMetricExporterOptions,
@@ -114,11 +179,15 @@ export class OTLPMetricExporterBase<
 {
   public _otlpExporter: T;
   private _aggregationTemporalitySelector: AggregationTemporalitySelector;
+  private _aggregationSelector: AggregationSelector;
 
   constructor(exporter: T, config?: OTLPMetricExporterOptions) {
     this._otlpExporter = exporter;
     this._aggregationTemporalitySelector = chooseTemporalitySelector(
       config?.temporalityPreference
+    );
+    this._aggregationSelector = chooseHistogramAggregation(
+      config?.histogramPreference
     );
   }
 
@@ -141,5 +210,9 @@ export class OTLPMetricExporterBase<
     instrumentType: InstrumentType
   ): AggregationTemporality {
     return this._aggregationTemporalitySelector(instrumentType);
+  }
+
+  selectAggregation(instrumentType: InstrumentType): Aggregation {
+    return this._aggregationSelector(instrumentType);
   }
 }