+++ packages/sdk-metrics/src/index.ts
@@ -24,6 +24,7 @@ export {
 export {
   AggregationSelector,
   AggregationTemporalitySelector,
+  DEFAULT_AGGREGATION_SELECTOR as DefaultAggregationSelector,
 } from './export/AggregationSelector';
 
 export { AggregationTemporality } from './export/AggregationTemporality';